chore: reorganize account
