--- src/core/client/account/fetches/index.ts
@@ -1,1 +1,0 @@
-export { default as CheckResetTokenFetch } from "./CheckResetTokenFetch";
