--- src/core/client/account/mutations/index.ts
@@ -1,1 +1,0 @@
-export { default as ResetPasswordMutation } from "./ResetPasswordMutation";
