+++ src/core/client/account/routeConfig.tsx
@@ -1,11 +1,10 @@
 import { makeRouteConfig, Route } from "found";
 import React from "react";
 
-import MainLayout from "./components/MainLayout";
 import ResetPassword from "./routes/password/reset/Index";
 
 export default makeRouteConfig(
-  <Route path="account" Component={MainLayout}>
+  <Route path="account">
     <Route path="password">
       <Route path="reset" {...ResetPassword.routeConfig} />
     </Route>
