+++ src/core/client/account/routes/NotFound.tsx
@@ -1,7 +1,7 @@
 import { HorizontalGutter, Typography } from "coral-ui/components";
 import React, { FunctionComponent } from "react";
 
-const NotFound: FunctionComponent = ({ children }) => (
+const NotFound: FunctionComponent = () => (
   <HorizontalGutter>
     <Typography variant="heading3">Not Found</Typography>
   </HorizontalGutter>
