+++ src/core/client/account/routes/password/reset/ResetTokenChecker.tsx
@@ -1,12 +1,11 @@
-import { Localized } from "fluent-react/compat";
-import React, { useEffect, useState } from "react";
-
-import { CheckResetTokenFetch } from "coral-account/fetches";
 import { ERROR_CODES } from "coral-common/errors";
 import { InvalidRequestError } from "coral-framework/lib/errors";
 import { useFetch } from "coral-framework/lib/relay";
 import { Delay, Flex, Spinner } from "coral-ui/components";
+import { Localized } from "fluent-react/compat";
+import React, { useEffect, useState } from "react";
 
+import CheckResetTokenFetch from "./CheckResetTokenFetch";
 import Sorry from "./Sorry";
 
 interface Props {
